+++ doc/changes.md
@@ -11,6 +11,7 @@
 * ADD: measurement auto-topup options (`topup`, `topupdays`, `topupsim` in the CLI)
 * ADD: ability to ask fo the backlog when streaming (`backlog`); automatically turned on for just scheduled one-offs
 * ADD: display authoritative section in native DNS output
+* ADD: support for aggregator client id (`clientid`)
 * FIX: reading results from a file failed
 * FIX: some errors were reported twice